+++ src/main/resources/db/test-data.sql
@@ -1,0 +1,1 @@
+insert into employee (id, name) values (1, 'A')
